Adding fixes for signedness warnings in liblangutil

+++ liblangutil/SourceReferenceFormatter.cpp
@@ -51,7 +51,7 @@ void SourceReferenceFormatter::printSourceLocation(SourceReference const& _ref)
 		);
 		m_stream << "^";
 		if (_ref.endColumn > _ref.startColumn + 2)
-			m_stream << string(_ref.endColumn - _ref.startColumn - 2, '-');
+			m_stream << string(static_cast<size_t>(_ref.endColumn - _ref.startColumn - 2), '-');
 		if (_ref.endColumn > _ref.startColumn + 1)
 			m_stream << "^";
 		m_stream << endl;
@@ -60,7 +60,7 @@ void SourceReferenceFormatter::printSourceLocation(SourceReference const& _ref)
 		m_stream <<
 			_ref.text <<
 			endl <<
-			string(_ref.startColumn, ' ') <<
+			string(static_cast<size_t>(_ref.startColumn), ' ') <<
 			"^ (Relevant source part starts here and spans across multiple lines)." <<
 			endl;
 }
